+++ src/backend/tokens.py
@@ -68,6 +68,10 @@ class Tokens:
         "{audio_channel_s_layout}", "Audio channel layout (L R C LFE Ls Rs Lb Rb)"
     )
     AUDIO_CODEC = FileToken("{audio_codec}", "Audio codec")
+    AUDIO_CODEC_NO_ATMOS = FileToken(
+        "{audio_codec_no_atmos}", "Audio codec with Atmos removed (TrueHD)"
+    )
+    ATMOS = FileToken("{atmos}", "Returns 'Atmos' if Atmos was detected")
     AUDIO_COMMERCIAL_NAME = FileToken(
         "{audio_commercial_name}", "Audio commercial name (Dolby Digital Plus)"
     )
@@ -211,6 +215,8 @@ class Tokens:
     )
 
     # NFO Tokens
+    MEDIA_TYPE = NfoToken("{media_type}", "Media type (Movie/Series)")
+    IS_ANIME = NfoToken("{is_anime}", "Returns 'Anime' if the release is anime")
     CHAPTER_TYPE = NfoToken(
         "{chapter_type}", "Chapter type (Named / Numbered (1 - 10) / Tagged)"
     )